toys/local_ci/requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
